% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apply_re288}
\alias{apply_re288}
\title{Apply delta run values}
\usage{
apply_re288(df, full_season_data)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Dataframe with run values
}
\description{
Apply delta run values
}
