% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_data_json}
\alias{get_data_json}
\title{Fetch live game JSON}
\usage{
get_data_json(ids_list)
}
\arguments{
\item{ids_list}{Integer vector of gamePk IDs.}
}
\value{
List of parsed JSON objects.
}
\description{
Fetch live game JSON
}
