% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_pbp_data}
\alias{get_pbp_data}
\title{Parse play-by-play}
\usage{
get_pbp_data(data_list)
}
\arguments{
\item{data_list}{List returned by \code{get_data_json()}.}
}
\value{
Tibble of pitch-by-pitch rows.
}
\description{
Parse play-by-play
}
