% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_pbp_player}
\alias{get_pbp_player}
\title{Get pbp data for a player}
\usage{
get_pbp_player(
  player_id,
  season = c(2025),
  start_date = NULL,
  end_date = NULL,
  game_type = c("R"),
  sport_id = c(1)
)
}
\arguments{
\item{player_id}{Integer MLBAM ID.}

\item{season}{Integer vector of years.}

\item{start_date, end_date}{Optional YYYY-MM-DD strings.}

\item{game_type}{Character vector (e.g., "R").}

\item{sport_id}{Integer.}
}
\value{
PBP data for a player
}
\description{
Get pbp data for a player
}
