% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_pbp_season}
\alias{get_pbp_season}
\title{Get pbp data for a season}
\usage{
get_pbp_season(
  season = c(2025),
  start_date = NULL,
  end_date = NULL,
  sport_id = c(1),
  game_type = c("R")
)
}
\arguments{
\item{season}{Integer vector of years.}

\item{start_date, end_date}{Optional YYYY-MM-DD strings.}

\item{sport_id}{Integer.}

\item{game_type}{Character vector (e.g., "R").}
}
\value{
PBP data for a season
}
\description{
Get pbp data for a season
}
