% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_player_games_list}
\alias{get_player_games_list}
\title{Get a player's game IDs}
\usage{
get_player_games_list(
  player_id,
  season = 2025,
  start_date = NULL,
  end_date = NULL,
  game_type = "R",
  sport_id = 1
)
}
\arguments{
\item{player_id}{Integer MLBAM ID.}

\item{season}{Integer vector of years.}

\item{start_date, end_date}{Optional YYYY-MM-DD strings.}

\item{game_type}{Character vector (e.g., "R").}

\item{sport_id}{Integer.}
}
\value{
Integer vector of gamePk IDs.
}
\description{
Get a player's game IDs
}
