% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_players}
\alias{get_players}
\title{Get players}
\usage{
get_players(sport_id = 1, season = 2025)
}
\arguments{
\item{sport_id}{Integer vector.}

\item{season}{Integer vector.}
}
\value{
Tibble of player info.
}
\description{
Get players
}
