% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_run_expectancy}
\alias{get_run_expectancy}
\title{Get run expectancy matrix}
\usage{
get_run_expectancy(df, matrix_type = c(24, 288))
}
\arguments{
\item{df}{Dataframe}

\item{matrix_type}{Integer either the re24 or re288 matrix}
}
\value{
run expectancy matrix
}
\description{
Get run expectancy matrix
}
