% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_schedule}
\alias{get_schedule}
\title{Get schedule}
\usage{
get_schedule(season = 2025, sport_id = 1, game_type = "R")
}
\arguments{
\item{season}{Integer vector of years.}

\item{sport_id}{Integer vector of sport IDs.}

\item{game_type}{Character vector of game type codes (e.g., "R").}
}
\value{
Data frame of schedule rows.
}
\description{
Get schedule
}
